Check model exists before checking if delete method exists in model.

+++ classes/model/temporal.php
@@ -636,7 +636,7 @@ class Model_Temporal extends Model
 				{
 					// Single relation, delete the model
 					$model = $rel->get($this);
-					if (method_exists($model, "delete")) {
+					if ($model && method_exists($model, "delete")) {
 						$model->delete($cascade);
 					}
 				}
